docs: 完成v0.0.76文档更新和发布准备
## 主要改进

### 文档系统升级
- 更新CLAUDE.md到v0.0.76（554行）
- 新增3个开发任务指南（添加Tauri命令、调试IPC、添加前端页面）
- 新增3个架构决策说明（Aho-Corasick、CAS、QueryExecutor）
- 新增4类性能基准数据和5个故障排查方案
- 移除过时的Kiro MCP Server调用指令

### 版本统一
- 更新所有文档版本号到0.0.76
- 同步README.md、CHANGELOG.md和模块级文档
- 更新.claude/index.json和任务清单

### 代码质量
- 修复Rust代码格式问题（4个文件）
- Clippy检查通过（零警告）
- TypeScript类型检查通过

### 发布准备
- 完成CAS迁移Phase 10所有任务（任务34-38）
- 生成完整的发布说明（RELEASE_NOTES_v0.0.76.md）
- 前端构建验证成功（851.28 kB, gzip: 259.21 kB）

## 详细变更

文档更新：
- CLAUDE.md: 从"做什么"升级为"怎么做"的完整开发指南
- README.md: 版本badge更新到0.0.76
- CHANGELOG.md: 添加[Unreleased]章节记录文档更新
- RELEASE_NOTES_v0.0.76.md: 创建详细的发布说明

代码格式化：
- src/search_engine/highlighting_engine.rs
- src/storage/integration_tests.rs
- src/storage/metadata_store.rs
- src/archive/gz_handler.rs

任务完成：
- ✅ 任务34: 代码审查（Clippy通过）
- ✅ 任务35: 生成变更日志
- ✅ 任务36: 更新版本号
- ✅ 任务37: 最终构建（前端构建成功）
- ✅ 任务38: 准备发布说明

影响：
- 文档实用性大幅提升，新开发者更容易上手
- 故障排查效率提升，有完整的排查流程
- 架构理解更深入，有关键决策说明和性能数据

+++ log-analyzer/src-tauri/src/storage/integration_tests.rs
@@ -89,7 +89,7 @@ mod tests {
         // Try to insert second file with same hash - CAS deduplication should return existing ID
         let file2 = FileMetadata {
             id: 0,
-            sha256_hash: hash2.clone(),  // 与 file1 相同的哈希
+            sha256_hash: hash2.clone(), // 与 file1 相同的哈希
             virtual_path: "logs/file2.log".to_string(),
             original_name: "file2.log".to_string(),
             size: content.len() as i64,
@@ -108,12 +108,22 @@ mod tests {
         );
 
         // 验证第一个文件仍然存在
-        let file1_id = metadata.get_file_by_virtual_path("logs/file1.log").await.unwrap().unwrap();
+        let file1_id = metadata
+            .get_file_by_virtual_path("logs/file1.log")
+            .await
+            .unwrap()
+            .unwrap();
         assert_eq!(file1_id.sha256_hash, hash1);
 
         // 验证第二个虚拟路径不存在（被 INSERT OR IGNORE 忽略）
-        let file2_result = metadata.get_file_by_virtual_path("logs/file2.log").await.unwrap();
-        assert!(file2_result.is_none(), "Second virtual path should not exist (ignored by UNIQUE constraint)");
+        let file2_result = metadata
+            .get_file_by_virtual_path("logs/file2.log")
+            .await
+            .unwrap();
+        assert!(
+            file2_result.is_none(),
+            "Second virtual path should not exist (ignored by UNIQUE constraint)"
+        );
 
         // Content should be stored only once in CAS
         let content_retrieved = cas.read_content(&file1_id.sha256_hash).await.unwrap();
@@ -121,7 +131,7 @@ mod tests {
 
         // Verify CAS deduplication worked - 两个哈希值相同，内容只存储一次
         assert!(cas.exists(&hash1));
-        assert!(cas.exists(&hash2));  // Same hash, so should exist
+        assert!(cas.exists(&hash2)); // Same hash, so should exist
     }
 
     #[tokio::test]
@@ -292,7 +302,7 @@ mod tests {
                 modified_time: 0,
                 mime_type: None,
                 parent_archive_id: None,
-                depth_level: i as i32,
+                depth_level: i,
             };
 
             metadata.insert_file(&file_meta).await.unwrap();
